requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
playwright==1.40.0
fake-useragent==1.4.0
requests==2.31.0
asyncio-mqtt==0.16.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
motor==3.3.2
pymongo==4.6.0
bcrypt==4.1.2